backend/requirements.txt
fastapi>=0.116.1,<1.0
uvicorn[standard]>=0.35.0,<1.0
supabase>=2.16.0,<3.0
python-jose[cryptography]>=3.5.0,<4.0
pydantic>=2.13.2,<3.0
pydantic-settings>=2.10.1,<3.0
pytest>=8.4.1,<9.0
httpx>=0.28.1,<1.0
